backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psutil==5.9.6
sqlalchemy==2.0.23
alembic==1.13.0
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
python-decouple==3.8
PyJWT==2.8.0
bcrypt==4.1.2
requests==2.31.0
